contact mail and statistics

+++ src/main/java/com/example/repositories/kadyaReprository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.entities.kadya;
 
-public interface kadyaReprository extends JpaRepository<kadya, Integer>{
+public interface kadyaReprository extends JpaRepository<kadya, Long>{
 	
 
 }
